requirements.txt
pyTelegramBotAPI
requests
python-dotenv
solana
solders
asyncio
